core_commands: move sample command to examplars
